% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ShinyGeode}
\alias{ShinyGeode}
\title{A user-friendly interface for accessing the spatial mapping and analysis functions of geode}
\usage{
ShinyGeode(...)
}
\description{
The ShinyGeode interface allows interactive visualization, mapping and analysis of geospatial
data through point-and-click, drop-down menus, checkboxes and sliders; no R coding is required.
The purpose of ShinyGeode is to provide a low barrier interface that benefits a wide range of
users. It allows rapid interaction with spatial data for identifying patterns, generating
hypotheses, and informing further investigation and analysis.
}
